initializing GUI classes & interfaces
