+++ .gitignore
@@ -3,6 +3,7 @@ out/
 !**/src/main/**/out/
 !**/src/test/**/out/
 .idea
+*.iml
 
 ### Eclipse ###
 .apt_generated
