+++ src/com/socialnetwork/connecthub/Main.java
@@ -1,0 +1,10 @@
+package com.socialnetwork.connecthub;
+
+import com.socialnetwork.connecthub.frontend.swing.view.LoginView;
+
+public class Main {
+    public static void main(String[] args) {
+
+    }
+
+}
