+++ src/com/socialnetwork/connecthub/frontend/swing/navigationhandler/NavigationHandler.java
@@ -1,0 +1,8 @@
+package com.socialnetwork.connecthub.frontend.swing.navigationhandler;
+
+import com.socialnetwork.connecthub.frontend.swing.view.View;
+
+public interface NavigationHandler {
+    View goToLoginView();
+    View goToSignUpView();
+}
