+++ src/com/socialnetwork/connecthub/frontend/swing/view/ContentCreationAreaView.java
@@ -1,0 +1,4 @@
+package com.socialnetwork.connecthub.frontend.swing.view;
+
+public class ContentCreationAreaView extends View {
+}
