+++ src/com/socialnetwork/connecthub/frontend/swing/view/MyProfileView.java
@@ -1,0 +1,4 @@
+package com.socialnetwork.connecthub.frontend.swing.view;
+
+public class MyProfileView extends View {
+}
