+++ src/com/socialnetwork/connecthub/frontend/swing/view/ProfileView.java
@@ -1,0 +1,4 @@
+package com.socialnetwork.connecthub.frontend.swing.view;
+
+public class ProfileView extends View {
+}
