+++ src/com/socialnetwork/connecthub/frontend/swing/view/SignUpView.java
@@ -1,0 +1,4 @@
+package com.socialnetwork.connecthub.frontend.swing.view;
+
+public class SignUpView extends View {
+}
