+++ src/com/socialnetwork/connecthub/frontend/swing/view/StoryView.java
@@ -1,0 +1,5 @@
+package com.socialnetwork.connecthub.frontend.swing.view;
+
+public class StoryView extends View {
+
+}
