+++ src/com/socialnetwork/connecthub/frontend/swing/view/View.java
@@ -1,0 +1,14 @@
+package com.socialnetwork.connecthub.frontend.swing.view;
+
+import javax.swing.*;
+
+public class View extends JFrame {
+    public View() {
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setSize(1500, 800);
+        setLocationRelativeTo(null);
+        setLayout(null);
+        setVisible(true);
+
+    }
+}
