--- src/java/com/socialnetwork/connecthub/frontend/swing/navigationHandler/NavigationHandler.java
@@ -1,7 +1,0 @@
-package java.com.socialnetwork.connecthub.frontend.swing.navigationHandler;
-
-import java.com.socialnetwork.connecthub.frontend.swing.view.View;
-
-public interface NavigationHandler {
-    View goToLoginView();
-}
