--- src/java/com/socialnetwork/connecthub/frontend/swing/view/LoginView.java
@@ -1,4 +1,0 @@
-package java.com.socialnetwork.connecthub.frontend.swing.view;
-
-public class LoginView extends View {
-}
