--- src/java/com/socialnetwork/connecthub/frontend/swing/view/NewsFeedView.java
@@ -1,4 +1,0 @@
-package java.com.socialnetwork.connecthub.frontend.swing.view;
-
-public class NewsFeedView extends View {
-}
