--- src/java/com/socialnetwork/connecthub/frontend/swing/view/StoryView.java
@@ -1,4 +1,0 @@
-package java.com.socialnetwork.connecthub.frontend.swing.view;
-
-public class StoryView extends View {
-}
