--- src/java/com/socialnetwork/connecthub/frontend/swing/view/View.java
@@ -1,6 +1,0 @@
-package java.com.socialnetwork.connecthub.frontend.swing.view;
-
-import javax.swing.*;
-
-public class View extends JFrame {
-}
