--- src/java/com/socialnetwork/connecthub/shared/Main.java
@@ -1,7 +1,0 @@
-package java.com.socialnetwork.connecthub.shared;
-
-public class Main {
-    public static void main(String[] args) {
-
-    }
-}
